+++ changelog.md
@@ -3,7 +3,7 @@
 ### v0.5.1
 
 1. Now we record sessions, with `start_time` and `end_time`, which would be imortant later on for statistical reasons. For this, `recents` table is altered, `session_id` column is added.
-2. Added countdown in days, with two config options `show_countdown` and `exam_date`. If enabled, it shows days remaining until `exam_date`.
+2. Added countdown in days, with two config options `show_eta_countdown` and `eta_date`. If enabled, it shows days remaining until `eta_date`.
 3. `create` command was not implemented until now :)
 4. Added aliases
    - `crawl` -> `crawler`.